Move widget utils into widget fodler
