+++ src/components/profile/widgets/widget-container.tsx
@@ -4,6 +4,7 @@ import { Widget } from './widget';
 import { NoWidgetsContent } from '@/components';
 import { AddWidgets } from '@/components/profile/widgets/add-widgets';
 import { DesktopOnlyAlert } from '@/components/profile/widgets/desktop-only-alert';
+import { parseWidgetTypeFromUrl } from '@/components/profile/widgets/util';
 import { useToast } from '@/components/ui/use-toast';
 import {
   useDeleteWidget,
@@ -21,7 +22,6 @@ import {
   WidgetSize,
   WidgetType,
 } from '@/types';
-import { parseWidgetTypeFromUrl } from '@/utils/icons';
 import { motion } from 'framer-motion';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import RGL, { Layout, WidthProvider } from 'react-grid-layout';
